infoschema: migrate cluster_tables_serial_test.go
